# api/openapi.yml
openapi: 3.0.3
info:
  title: Fansworld Creator Platform API
  version: 1.0.0
  description: |
    Production-ready API for creator subscription platform with OnlyFans-level functionality.
    Supports content creation, subscription management, messaging, and creator monetization.
  contact:
    name: Fansworld API Team
    email: api@fansworld.com
  license:
    name: Proprietary
    url: https://fansworld.com/license
servers:
  - url: https://api.fansworld.com/v1
    description: Production server
  - url: https://staging-api.fansworld.com/v1
    description: Staging server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /conversations:
    get:
      summary: Get user conversations
      tags: [Messaging]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Conversations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationsResponse'

    post:
      summary: Create new conversation
      tags: [Messaging]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId]
              properties:
                participantId: { type: string }
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
  /conversations/{conversationId}/messages:
    get:
      summary: Get conversation messages
      tags: [Messaging]
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: before
          schema:
            type: string
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'

    post:
      summary: Send message
      tags: [Messaging]
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '402':
          description: Payment required for PPV message

  # Search & Discovery
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PublicUser:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, format: uri }
        verified: { type: boolean }
...
    Conversation:
      type: object
      properties:
        id: { type: string }
        participants: { type: array, items: { $ref: '#/components/schemas/PublicUser' } }
        lastMessage: { $ref: '#/components/schemas/Message' }
        unreadCount: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
    ConversationsResponse:
      type: object
      properties:
        conversations: { type: array, items: { $ref: '#/components/schemas/Conversation' } }
        nextCursor: { type: string }
        hasMore: { type: boolean }
    Message:
      type: object
      properties:
        id: { type: string }
        conversationId: { type: string }
        senderId: { type: string }
        content: { type: string }
        mediaUrls: { type: array, items: { type: string, format: uri } }
        isLocked: { type: boolean }
        price: { type: number, minimum: 0 }
        hasAccess: { type: boolean }
        sentAt: { type: string, format: date-time }
    MessagesResponse:
      type: object
      properties:
        messages: { type: array, items: { $ref: '#/components/schemas/Message' } }
        hasMore: { type: boolean }
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, maxLength: 1000 }
        mediaIds: { type: array, items: { type: string } }
        isLocked: { type: boolean, default: false }
        price: { type: number, minimum: 0 }

    # Search Schemas
...